+++ src/parser/parse.ts
@@ -217,6 +217,7 @@ export class Parser<T extends ParserInput, TFlags extends OutputFlags<T['flags']
       const flag = this.input.flags[token.flag]
 
       if (!flag) throw new CLIError(`Unexpected flag ${token.flag}`)
+
       if (flag.type === 'boolean') {
         if (token.input === `--no-${flag.name}`) {
           flags[token.flag] = false
@@ -256,7 +257,7 @@ export class Parser<T extends ParserInput, TFlags extends OutputFlags<T['flags']
     for (const k of Object.keys(this.input.flags)) {
       const flag = this.input.flags[k]
       if (flags[k]) continue
-      if (flag.env && Object.prototype.hasOwnProperty.call(process.env, flag.env)) {
+      if (flag.env && Reflect.has(process.env, flag.env)) {
         const input = process.env[flag.env]
         if (flag.type === 'option') {
           if (input) {
@@ -271,9 +272,27 @@ export class Parser<T extends ParserInput, TFlags extends OutputFlags<T['flags']
       }
 
       if (!(k in flags) && flag.default !== undefined) {
-        this.metaData.flags[k] = {setFromDefault: true}
-        const defaultValue = (typeof flag.default === 'function' ? await flag.default({options: flag, flags, ...this.context}) : flag.default)
+        this.metaData.flags[k] = {...this.metaData.flags[k], setFromDefault: true}
+        const defaultValue = (typeof flag.default === 'function' ? await flag.default({
+          options: flag,
+          flags, ...this.context,
+        }) : flag.default)
         flags[k] = defaultValue
+      }
+    }
+
+    for (const k of Object.keys(this.input.flags)) {
+      if ((k in flags) && Reflect.has(this.input.flags[k], 'defaultHelp')) {
+        try {
+          const defaultHelpProperty = Reflect.get(this.input.flags[k], 'defaultHelp')
+          const defaultHelp = (typeof defaultHelpProperty === 'function' ? await defaultHelpProperty({
+            options: flags[k],
+            flags, ...this.context,
+          }) : defaultHelpProperty)
+          this.metaData.flags[k] = {...this.metaData.flags[k], defaultHelp}
+        } catch {
+          // no-op
+        }
       }
     }