+++ test/command/command.test.ts
@@ -103,7 +103,7 @@ describe('command', () => {
       }
       }
 
-      const c = await toCached(C)
+      const c = await toCached(C, undefined, false)
 
       expect(c).to.deep.equal({
         id: 'foo:bar',
@@ -448,5 +448,124 @@ describe('command', () => {
     })
     .do(ctx => expect(ctx.stdout).to.equal('json output: {"a":"foobar"}\n'))
     .it('test stdout EPIPE swallowed')
+  })
+  describe('json enabled and pass-through tests', () => {
+    fancy
+    .stdout()
+    .do(async () => {
+      class CMD extends Command {
+        static enableJsonFlag = true
+        async run() {
+          this.log('not json output')
+        }
+      }
+      const cmd = new CMD([], {} as any)
+      expect(cmd.jsonEnabled()).to.equal(false)
+    })
+    .it('json enabled/pass through disabled/no --json flag/jsonEnabled() should be false')
+
+    fancy
+    .stdout()
+    .do(async () => {
+      class CMD extends Command {
+        static enableJsonFlag = true
+        async run() {}
+      }
+      const cmd = new CMD(['--json'], {} as any)
+      expect(cmd.jsonEnabled()).to.equal(true)
+    })
+    .it('json enabled/pass through disabled/--json flag before --/jsonEnabled() should be true')
+
+    fancy
+    .stdout()
+    .do(async () => {
+      class CMD extends Command {
+          static enableJsonFlag = true
+          static '--' = true
+          async run() {
+            const {flags} = await cmd.parse(CMD, ['--json'])
+            expect(flags.json).to.equal(true, 'json flag should be true')
+          }
+      }
+      const cmd = new CMD(['--json'], {} as any)
+      expect(cmd.jsonEnabled()).to.equal(true)
+    })
+    .it('json enabled/pass through enabled/--json flag before --/jsonEnabled() should be true')
+
+    fancy
+    .stdout()
+    .do(async () => {
+      class CMD extends Command {
+          static enableJsonFlag = true
+          static '--' = true
+          async run() {
+            const {flags} = await cmd.parse(CMD, ['--', '--json'])
+            expect(flags.json).to.equal(false, 'json flag should be false')
+            expect(this.passThroughEnabled).to.equal(true, 'pass through should be true')
+          }
+      }
+      const cmd = new CMD(['--', '--json'], {} as any)
+      expect(cmd.jsonEnabled()).to.equal(false)
+    })
+    .it('json enabled/pass through enabled/--json flag after --/jsonEnabled() should be false')
+
+    fancy
+    .stdout()
+    .do(async () => {
+      class CMD extends Command {
+          static enableJsonFlag = true
+          static '--' = true
+          async run() {
+            const {flags} = await cmd.parse(CMD, ['--foo', '--json'])
+            expect(flags.json).to.equal(true, 'json flag should be true')
+          }
+      }
+      const cmd = new CMD(['foo', '--json'], {} as any)
+      expect(cmd.jsonEnabled()).to.equal(true)
+    })
+    .it('json enabled/pass through enabled/--json flag before --/extra param/jsonEnabled() should be true')
+
+    fancy
+    .stdout()
+    .do(async () => {
+      class CMD extends Command {
+          static enableJsonFlag = true
+          static '--' = true
+          async run() {
+            const {flags} = await cmd.parse(CMD, ['--foo', '--', '--json'])
+            expect(flags.json).to.equal(false, 'json flag should be false')
+            expect(this.passThroughEnabled).to.equal(true, 'pass through should be true')
+          }
+      }
+      const cmd = new CMD(['--foo', '--', '--json'], {} as any)
+      expect(cmd.jsonEnabled()).to.equal(false)
+    })
+    .it('json enabled/pass through enabled/--json flag after --/extra param/jsonEnabled() should be false')
+
+    fancy
+    .stdout()
+    .do(async () => {
+      class CMD extends Command {
+          static enableJsonFlag = true
+          static '--' = true
+          async run() {}
+      }
+      const cmd = new CMD(['--json', '--'], {} as any)
+      expect(cmd.jsonEnabled()).to.equal(true)
+    })
+    .it('json enabled/pass through enabled/--json flag before --/jsonEnabled() should be true')
+
+    fancy
+    .stdout()
+    .do(async () => {
+      class CMD extends Command {
+          static enableJsonFlag = false
+          static '--' = true
+          async run() {}
+      }
+      const cmd = new CMD(['--json'], {} as any)
+      expect(cmd.jsonEnabled()).to.equal(false)
+    })
+    .it('json disabled/pass through enable/--json flag before --/jsonEnabled() should be false')
   })
 })